Move split_buffers() to helpers
